companies/wpflow/requirements.txt
mcp>=1.2.0,<2.0.0
httpx>=0.27.0,<1.0.0
pydantic>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
